refactor: split Flex folder creation modules
