+++ docs/plans/codebase-maintenance-roadmap.md
@@ -38,6 +38,7 @@ This roadmap turns the current maintenance audit into concrete, sequenced backlo
 - 2026-05-16: Phase 2 completed across P2-01 through P2-05, retaining the already-complete P2-04 permission layer. Component/page direct Supabase call hits were reduced from the post-Phase-1 scan of `318` to `12`; inline `queryKey: [...]` property arrays were replaced with `queryKeys` factories and now scan at `0`; repeated tour-management mutation toast/invalidation boilerplate now uses `useMutationFeedback`; wallboard calendar windows use Europe/Madrid date-key utilities; and timezone tests cover Madrid spring and autumn DST boundaries. Local validation passed with `npm run typecheck`, `npm run lint`, `npm run test:run`, `npm run build`, `npm run cap:sync`, and `git diff --check`.
 - 2026-05-16: P3-01 completed by splitting `src/components/technician/DetailsModal.tsx` from `2050` lines into a `79` line shell plus `src/components/technician/details-modal/` modules for modal data loading, document actions, formatting helpers, shared types, and tab sections. Added focused regression tests for formatter behavior, document open/download actions, and the Info/Documents tab render paths. Local validation passed with `npm run typecheck`, targeted Vitest coverage, `npm run lint`, `npm run test:run`, `npm run build`, `npm run cap:sync`, and `git diff --check`.
 - 2026-05-16: P3-02 completed by splitting `src/components/jobs/cards/JobCardActions.tsx` from `1858` lines into a `210` line shell plus `src/components/jobs/cards/job-card-actions/` modules for production/warehouse WhatsApp, Flex opening and selector hosting, technical power packs, shared action formatting, action buttons, and dialogs. Existing Flex and technical-power tests remain in place, and focused smoke coverage now covers transport, document/upload actions, warehouse WhatsApp, and production WhatsApp dialog generation. Local validation passed with `npm run typecheck`, targeted Vitest coverage, `npm run lint`, `npm run test:critical`, `npm run test:run`, `npm run build`, `npm run cap:sync`, `npm run test:e2e`, and `git diff --check`.
+- 2026-05-16: P3-03 completed by splitting `src/utils/flex-folders/folders.ts` from `1631` lines into a `1` line compatibility entrypoint plus `src/utils/flex-folders/folder-creation/` modules for orchestration, dryhire creation and cleanup, tourdate creation, standard job creation, commercial extras, shared helpers, and shared creation types. The new creation boundary also uses a Flex-folder-specific job type instead of preserving the broad legacy `any`. Local validation passed with `npm run typecheck`, targeted Flex folder Vitest coverage, `npm run lint`, `npm run test:critical`, `npm run test:run`, `npm run build`, `npm run cap:sync`, `npm run test:e2e`, and `git diff --check`.
 
 ## Phase 0: Safety and Type Baseline
